move code out of platform-operator/util package
