+++ platform-operator/apis/verrazzano/v1alpha1/validate.go
@@ -12,7 +12,7 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/util/semver"
+	"github.com/verrazzano/verrazzano/platform-operator/internal/semver"
 
 	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
