+++ platform-operator/apis/verrazzano/v1alpha1/validate_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/util/semver"
+	"github.com/verrazzano/verrazzano/platform-operator/internal/semver"
 
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
