+++ platform-operator/internal/helm/testdata/Chart.yaml
