+++ platform-operator/internal/k8s/certificate/certificate.go
