+++ platform-operator/internal/k8s/certificate/certificate_test.go
