+++ platform-operator/internal/k8s/netpolicy/netpolicy.go
