+++ platform-operator/internal/k8s/netpolicy/netpolicy_test.go
