+++ platform-operator/internal/os/exec_test.go
