+++ platform-operator/internal/semver/semver_test.go
